sched/fair: Reduce redundancy in imbalance tasks calculation
It's not needed to use right shift operator for both conditions when we can just
generalize it later only 1 time. This is in preparation to merge commit "sched/fair:
Implement prefer sibling imbalance calculation between asymmetric group".

Intentionally, the right shift operator does not cover the case for NUMA balancing
here to retain the consistency of this patch since NUMA balancing uses the total
imbalance whereas load balancing uses the number of tasks that must move to restore
imbalance.

This alters no functionality and is extracted from upstream [1].

[1] https://github.com/torvalds/linux/commit/cb29a5c19d2d68a

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>